@@ -8,50 +8,52 @@
     "build": "ng build --prod",
     "test": "ng test",
     "lint": "ng lint",
-    "e2e": "ng e2e"
+    "e2e": "ng e2e --port=4202"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "6.0.0",
-    "@angular/cdk": "6.0.1",
-    "@angular/common": "6.0.0",
-    "@angular/compiler": "6.0.0",
-    "@angular/core": "6.0.0",
-    "@angular/flex-layout": "6.0.0-beta.15",
-    "@angular/forms": "6.0.0",
-    "@angular/http": "6.0.0",
-    "@angular/material": "6.0.1",
-    "@angular/platform-browser": "6.0.0",
-    "@angular/platform-browser-dynamic": "6.0.0",
-    "@angular/router": "6.0.0",
-    "@angular/service-worker": "6.0.0",
-    "@ngxs/store": "3.0.1",
+    "@angular/animations": "6.1.9",
+    "@angular/cdk": "6.4.7",
+    "@angular/common": "6.1.9",
+    "@angular/compiler": "6.1.9",
+    "@angular/core": "6.1.9",
+    "@angular/flex-layout": "6.0.0-beta.18",
+    "@angular/forms": "6.1.9",
+    "@angular/http": "6.1.9",
+    "@angular/material": "6.4.7",
+    "@angular/platform-browser": "6.1.9",
+    "@angular/platform-browser-dynamic": "6.1.9",
+    "@angular/router": "6.1.9",
+    "@angular/service-worker": "6.1.9",
+    "@ngxs/store": "3.2.0",
     "core-js": "2.5.6",
-    "firebase": "5.0.1",
-    "idb-keyval": "3.0.3",
-    "rxjs": "6.1.0",
+    "firebase": "5.5.3",
+    "idb-keyval": "3.1.0",
+    "rxjs": "6.3.3",
     "zone.js": "0.8.26"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "0.6.0",
-    "@angular/cli": "6.0.0",
-    "@angular/compiler-cli": "6.0.0",
-    "@angular/language-service": "6.0.0",
-    "@types/jasmine": "2.8.7",
-    "@types/jasminewd2": "2.0.3",
-    "@types/node": "9.6.13",
-    "codelyzer": "4.3.0",
-    "jasmine-core": "3.1.0",
+    "@angular-devkit/build-angular": "0.8.4",
+    "@angular/cli": "6.2.4",
+    "@angular/compiler-cli": "6.1.9",
+    "@angular/language-service": "6.1.9",
+    "@netbasal/spectator": "2.0.2",
+    "@types/jasmine": "2.8.9",
+    "@types/jasminewd2": "2.0.5",
+    "@types/node": "10.11.4",
+    "codelyzer": "4.5.0",
+    "jasmine-core": "3.2.1",
     "jasmine-spec-reporter": "4.2.1",
-    "karma": "2.0.2",
+    "karma": "3.0.0",
     "karma-chrome-launcher": "2.2.0",
-    "karma-coverage-istanbul-reporter": "1.4.2",
+    "karma-coverage-istanbul-reporter": "2.0.4",
     "karma-jasmine": "1.1.2",
-    "karma-jasmine-html-reporter": "1.0.0",
-    "prettier": "1.12.1",
-    "protractor": "5.3.1",
-    "ts-node": "6.0.3",
-    "tslint": "5.10.0",
-    "typescript": "2.7.2"
+    "karma-jasmine-html-reporter": "1.3.1",
+    "prettier": "1.14.3",
+    "protractor": "5.4.1",
+    "spectator": "0.0.4",
+    "ts-node": "7.0.1",
+    "tslint": "5.11.0",
+    "typescript": "2.9.2"
   }
 }
