@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 
-const routes: Routes = [{ path: '', loadChildren: 'app/module/home/home.module#HomeModule' }];
+const routes: Routes = [{ path: '', loadChildren: 'src/app/module/home/home.module#HomeModule' }];
 
 @NgModule({
   imports: [RouterModule.forRoot(routes)],
