@@ -1,18 +1,14 @@
-import { TestBed, async } from '@angular/core/testing';
-import { RouterTestingModule } from '@angular/router/testing';
+import { createTestComponentFactory } from '@netbasal/spectator';
 
 import { AppComponent } from './app.component';
 
 describe('AppComponent', () => {
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      imports: [RouterTestingModule],
-      declarations: [AppComponent]
-    }).compileComponents();
-  }));
-  it('should create the app', async(() => {
-    const fixture = TestBed.createComponent(AppComponent);
-    const app = fixture.debugElement.componentInstance;
-    expect(app).toBeTruthy();
-  }));
+  const createComponent = createTestComponentFactory({
+    component: AppComponent,
+    shallow: true
+  });
+
+  it('should successfully instantiate', () => {
+    expect(createComponent().element).toExist();
+  });
 });
