@@ -2,5 +2,5 @@ import { DocumentSnapshot, SnapshotOptions } from '@firebase/firestore-types';
 
 export interface DocumentTypedSnapshot<T> extends DocumentSnapshot {
   data(options?: SnapshotOptions): T;
-  get(fieldPath: keyof T, options?: SnapshotOptions): any;
+  get(fieldPath: string, options?: SnapshotOptions): T[keyof T];
 }
