@@ -1,37 +1,29 @@
+import firebase from 'firebase/app';
+import 'firebase/auth';
 import 'firebase/firestore';
-import { firestore, initializeApp } from 'firebase';
 
-import { NgModule } from '@angular/core';
+import { NgModule, NgZone } from '@angular/core';
 
-import { FirebaseConfig } from './interface/firebase-config.interface';
 import { FirebaseService } from './firebase.service';
+import { environment } from '../../../../environments/environment';
 
-export function _firebaseAppFactory(config: FirebaseConfig): FirebaseService {
-  const fs: FirebaseService = initializeApp(config);
+export function _firebaseAppFactory(ngZone: NgZone): FirebaseService {
+  const fs: FirebaseService = ngZone.runOutsideAngular(() => firebase.initializeApp(environment.firebase));
 
-  firestore().settings({
+  firebase.firestore().settings({
     timestampsInSnapshots: true
   });
 
   return fs;
 }
 
-@NgModule({})
-export class FirebaseModule {
-  static initialize(config: FirebaseConfig) {
-    return {
-      ngModule: FirebaseModule,
-      providers: [
-        {
-          provide: FirebaseConfig,
-          useValue: config
-        },
-        {
-          provide: FirebaseService,
-          useFactory: _firebaseAppFactory,
-          deps: [FirebaseConfig]
-        }
-      ]
-    };
-  }
-}
+@NgModule({
+  providers: [
+    {
+      provide: FirebaseService,
+      useFactory: _firebaseAppFactory,
+      deps: [NgZone]
+    }
+  ]
+})
+export class FirebaseModule {}
