@@ -1,8 +1,7 @@
-import { auth } from 'firebase';
+import firebase from 'firebase/app';
+import { User } from '@firebase/auth-types';
 
 import { Injectable } from '@angular/core';
-
-import { User } from '@firebase/auth-types';
 
 import { Store } from '@ngxs/store';
 
@@ -28,7 +27,7 @@ export class AuthStateService {
   }
 
   public initAuth(): Promise<void> {
-    return this.store.dispatch(new SetPersistence(auth.Auth.Persistence.LOCAL)).toPromise();
+    return this.store.dispatch(new SetPersistence(firebase.auth.Auth.Persistence.LOCAL)).toPromise();
   }
 
   private updateUser(user: User | null): void {
