@@ -1,6 +1,6 @@
-import { Observable } from 'rxjs/index';
+import { Observable } from 'rxjs';
 
-import { Component } from '@angular/core';
+import { ChangeDetectionStrategy, Component } from '@angular/core';
 
 import { Select } from '@ngxs/store';
 
@@ -8,7 +8,8 @@ import { AuthState } from '../../../../shared/states/auth.state';
 
 @Component({
   templateUrl: './home.component.html',
-  styleUrls: ['./home.component.scss']
+  styleUrls: ['./home.component.scss'],
+  changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class HomeComponent {
   @Select(AuthState.loggedIn) readonly loggedIn$!: Observable<boolean>;
