@@ -1,26 +1,22 @@
-import { Component, Input } from '@angular/core';
+import { ChangeDetectionStrategy, Component, Input } from '@angular/core';
 
 import { Store } from '@ngxs/store';
 
-import { MediaDevicesService } from '../../../../core/services/media-devices.service';
+import { RemoveDevice } from '../../../../shared/actions/devices.action';
 import { LocalDeviceModel } from '../../../../shared/states/devices.state';
 import { DocumentTypedSnapshot } from '../../../../core/interface/document-data.interface';
-import { RemoveDevice } from '../../../../shared/actions/devices.action';
 import { StreamConnectionService } from '../../../../core/services/stream-connection.service';
 
 @Component({
   selector: 'vc-stream-thumb',
   templateUrl: './stream-thumb.component.html',
-  styleUrls: ['./stream-thumb.component.scss']
+  styleUrls: ['./stream-thumb.component.scss'],
+  changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class StreamThumbComponent {
   @Input() device?: DocumentTypedSnapshot<LocalDeviceModel>;
 
-  constructor(
-    private readonly md: MediaDevicesService,
-    private readonly store: Store,
-    private readonly sc: StreamConnectionService
-  ) {}
+  constructor(private readonly store: Store, private readonly sc: StreamConnectionService) {}
 
   public onDeleteDevice(): void {
     if (this.device) {
