@@ -1,4 +1,4 @@
-import { NgModule} from '@angular/core';
+import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { CommonModule } from '@angular/common';
 import { FlexLayoutModule } from '@angular/flex-layout';
@@ -8,7 +8,8 @@ import {
   MatFormFieldModule,
   MatIconModule,
   MatInputModule,
-  MatRadioModule, MatToolbarModule
+  MatRadioModule,
+  MatToolbarModule
 } from '@angular/material';
 
 import { SafeStylePipe } from './pipes/safe-style.pipe';
@@ -18,7 +19,7 @@ import { AvatarComponent } from './components/avatar/avatar.component';
 import { LoginDialogComponent } from './components/dialog/login-dialog/login-dialog.component';
 import { AddDeviceDialogComponent } from './components/dialog/add-device-dialog/add-device-dialog.component';
 import { StreamViewComponent } from './components/stream-view/stream-view.component';
-import {DialogMoverDirective} from "./directives/dialog-mover.directive";
+import { DialogMoverDirective } from './directives/dialog-mover.directive';
 
 @NgModule({
   imports: [
