@@ -1,9 +1,10 @@
-import { Component, OnInit } from '@angular/core';
+import { ChangeDetectionStrategy, Component, OnInit } from '@angular/core';
 
 @Component({
   selector: 'vc-app-footer',
   templateUrl: './app-footer.component.html',
-  styleUrls: ['./app-footer.component.scss']
+  styleUrls: ['./app-footer.component.scss'],
+  changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class AppFooterComponent implements OnInit {
   constructor() {}
